backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
flask-sqlalchemy==3.1.1
pillow==10.1.0
numpy==1.26.2
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
torch==2.1.1
torchvision==0.16.1
scikit-image==0.22.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
werkzeug==3.0.1
redis==5.0.1
celery==5.3.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.5.2
pyjwt==2.8.0
bcrypt==4.1.2
python-multipart==0.0.6
aiofiles==23.2.1
prometheus-client==0.19.0
sentry-sdk[flask]==1.39.1
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
black==23.12.1
flake8==6.1.0
mypy==1.7.1
